로그인시 위도, 경도 정보 세션에 저장 로그아웃시 위도, 경도 정보 세션에서 삭제
운동시설 전체 리스트에서 거리별 상품 리스트 구현(현재 위치에서 5km 이내 상품만)

+++ picktness/src/main/java/com/pick/member/vo/MemberVO.java
@@ -22,8 +22,8 @@ public class MemberVO {
 	private String phoneConsent;
 	private String del_yn;
 	private Date joinDate;
-	private BigDecimal lat;
-	private BigDecimal lng;
+	private double lat;
+	private double lng;
 	public String getId() {
 		return id;
 	}
@@ -120,17 +120,18 @@ public class MemberVO {
 	public void setJoinDate(Date joinDate) {
 		this.joinDate = joinDate;
 	}
-	public BigDecimal getLat() {
+	public double getLat() {
 		return lat;
 	}
-	public void setLat(BigDecimal lat) {
+	public void setLat(double lat) {
 		this.lat = lat;
 	}
-	public BigDecimal getLng() {
+	public double getLng() {
 		return lng;
 	}
-	public void setLng(BigDecimal lng) {
+	public void setLng(double lng) {
 		this.lng = lng;
 	}
-	
+
+
 }
